Changes
This version contain the adrs settings, the min and max frequency and a
plot for the adrs.
The data gris disposition was changed to allow big number of rows on the
data. The continuous and discrete sound was allowed.
A text control to enter some predefined functionalities was added.
Now the comunication with octave is made from the same element of the
functionalities. The user can enter a number to jump on the abssisa and
tempo sliders.