solving nan issues
I detect an error at the end of the slider, the solution was set the max
slider value to 'x.size-1' instead of 'x.size'.
I use the np.nanmin and np.nanmax functions to ignore the nan values on
the array when the max and min values are computed to plot the red line
and set the limits of the selected value on the slider.
Erase some unnecessary print statements.