Documentation
This part include the documentacion of the methods to plot data. In
addition, the order of replot was change to not hide the marks under the
data plot, and discard marks made when the marker is at the end of the
data.
No, when the user press mark point while the reproduction is paused, the
black mark appear above the red line of the position.