Add cluster energy param
To relate the cluster energy to the volume, first test the volume limits
and conclude a change of amplitud from 0 to 2000 on the sine wave. So
change all the amplitud default values to 2000.
Second revise the min and max value of the transverse energy of the
cluster (third number of the cluster array), and stablish the range
between [0;100].
Finally, add this to the cluster sound and plot. First, normalize the
cluster energy dividing the value by 100, and then: for sound, multiply
that by 2000 and set an offset of 100 when the min energy is not 0; for
plot, multiply the normalize energy by 10 and set an offset value of 2
when the min energy is not 0.